Refactoring.

+++ modules/core/client/microcontrollers/microcontrollers-list/microcontrollers.client.controller.js
@@ -2,19 +2,19 @@
 
 angular.module('core')
     .controller('MicrocontrollersController',
-        function ($scope, microcontrollersModel, $uibModal, placesModel) {
+        function($scope, microcontrollersModel, $uibModal, placesModel, confirmation) {
             $scope.microcontrollersModel = microcontrollersModel.model;
             $scope.searchText = '';
-            $scope.addEdit = function (microcontrollerId) {
-                placesModel.load().then(function () {
+            $scope.addEdit = function(microcontrollerId) {
+                placesModel.load().then(function() {
                     var modalInstance = $uibModal.open({
                         templateUrl: 'modules/core/client/microcontrollers/microcontroller-form/microcontrollers.add-edit.client.view.html',
                         controller: 'AddEditMicrocontrollerController',
                         resolve: {
-                            data: function () {
+                            data: function() {
                                 return {
                                     modelId: microcontrollerId,
-                                    places: placesModel.model.places.map(function (p) {
+                                    places: placesModel.model.places.map(function(p) {
                                         return {
                                             id: p.id,
                                             title: p.title
@@ -29,17 +29,19 @@ angular.module('core')
                 });
             };
 
-            $scope.delete = function (id) {
-                microcontrollersModel.delete(id);
-            };
-
-
-            $scope.bulkDelete = microcontrollersModel.bulkDelete;
-            $scope.bulkDeleteDisabled = function () {
-                return !microcontrollersModel.model.microcontrollers.some(function (microcontroller) {
-                    return microcontroller.isSelected;
+            $scope.delete = function(id) {
+                confirmation.confirm('Warning!', 'Do you really want to delete this item?', function() {
+                    microcontrollersModel.delete(id);
                 });
             };
 
-            microcontrollersModel.load();
+            $scope.bulkDelete = function() {
+                confirmation.confirm('Warning!', 'Do you really want to delete these items?', function() {
+                    microcontrollersModel.bulkDelete();
+                });
+            };
+
+            $scope.bulkDeleteDisabled = microcontrollersModel.bulkDeleteDisabled;
+
+            $scope.load = microcontrollersModel.load;
         });
